# Display related optimizations and performance improvements 