- Reduce screen writes by preventing unnecessary clear_screen_region activity.
- Do not perform Frame Buffer updates when nothing is changing.
- Don't pad test descriptions.