@debrouxl 2025-02-09
I was wondering whether drawing functions slowed down memtest86+, especially for short-lived tests such as test 0, but I never tried to investigate.
Speed-wise, it certainly makes sense to add shortcuts to `lfb*_put_char()` functions, as these shortcuts are probably taken frequently enough to offset their cost. 

The third patch undoes an earlier change of mine. On x86_64, switching to fixed-size strings replaced (8 bytes of pointer to string data + 24 bytes of relocation + > 8 bytes of string data) by a constant 40 bytes of string data, so it was, and still is, a clear optimization in terms of raw bytes: on my computer, in `memtest_shared` (which contains sections not present into the final binary, but still), the end of `.rela.got` is at offset 1f9b0 instead of 1fc48 with the third patch:
```
(baseline)
  0 .text         00016c49  0000000000000000  0000000000000000  00001000  2**5
                  CONTENTS, ALLOC, LOAD, READONLY, CODE
  1 .rodata       000039f0  0000000000016c60  0000000000016c60  00017c60  2**5
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  2 .dynsym       000019b0  000000000001a650  000000000001a650  0001b650  2**3
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  3 .dynstr       00000e16  000000000001c000  000000000001c000  0001d000  2**0
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  4 .gnu.hash     00000970  000000000001ce18  000000000001ce18  0001de18  2**3
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  5 .dynamic      00000100  000000000001d788  000000000001d788  0001e788  2**3
                  CONTENTS, ALLOC, LOAD, DATA
  6 .rela.data    00001128  000000000001d888  000000000001d888  0001e888  2**3
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  7 .rela.got     000000a8  000000000001e9b0  000000000001e9b0  0001f9b0  2**3
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  8 .data         00006b28  000000000001f000  000000000001f000  00020000  2**12
                  CONTENTS, ALLOC, LOAD, DATA
  9 .got          00000050  0000000000025b28  0000000000025b28  00026b28  2**3
                  CONTENTS, ALLOC, LOAD, DATA
 10 .bss          0004a380  0000000000025b80  0000000000025b80  00026b78  2**5
                  ALLOC

(with patch 3 only)
  0 .text         00016c70  0000000000000000  0000000000000000  00001000  2**5
                  CONTENTS, ALLOC, LOAD, READONLY, CODE
  1 .rodata       00003b5a  0000000000016c80  0000000000016c80  00017c80  2**5
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  2 .dynsym       000019b0  000000000001a7e0  000000000001a7e0  0001b7e0  2**3
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  3 .dynstr       00000e16  000000000001c190  000000000001c190  0001d190  2**0
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  4 .gnu.hash     00000970  000000000001cfa8  000000000001cfa8  0001dfa8  2**3
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  5 .dynamic      00000100  000000000001d918  000000000001d918  0001e918  2**3
                  CONTENTS, ALLOC, LOAD, DATA
  6 .rela.data    00001230  000000000001da18  000000000001da18  0001ea18  2**3
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  7 .rela.got     000000a8  000000000001ec48  000000000001ec48  0001fc48  2**3
                  CONTENTS, ALLOC, LOAD, READONLY, DATA
  8 .data         000069c8  000000000001f000  000000000001f000  00020000  2**12
                  CONTENTS, ALLOC, LOAD, DATA
  9 .got          00000050  00000000000259c8  00000000000259c8  000269c8  2**3
                  CONTENTS, ALLOC, LOAD, DATA
 10 .bss          0004a2e0  0000000000025a20  0000000000025a20  00026a18  2**5
                  ALLOC
```
The 4K alignment constraint for .data currently (but certainly not back then) nullifies the advantage of the fixed-size strings into the final binary. Let the binary grow a bit, in such a way that the end of `.rela.got` spills beyond 0x20000 when using string pointers but not when using fixed-size strings, and the change in patch 3 temporarily raises the size by 4K... until the program grows further and string pointer + relocation + string data saves a bit of space again.
The optimal representation effectively depends on the rest of the program :)