@x86fr 2025-02-09
PS: I've posted a recap of previous discussions about code size and cpu cycle counting in #492