@01e3 2025-02-09
Oh, I was not aware that a "constant 40 bytes of string data" was a recent change and done intentionally, and I assumed it had always been like this - since the very, very early days. I should have checked, my bad.

Now, while I find "regular" text strings easier to use and maintain, we should not expect this section of the code to change frequently. I assume I should drop it then, especially under the spirit of Sam's guideline https://github.com/memtest86plus/memtest86plus/discussions/492?

What about the two other commits?