@x86fr 2025-02-09
Adding a new definition for `SCREEN_MAX_COL`, well, why not.
Adding another level of nested macros with `printf_region()` is only a matter of code readability. I'm not sure if that many level of macro/function nested is really better, but that's trivial.
About returning if char in FB is identical, it saves a few CPU cycles, but can add potential issues (or not) with the serial TTY code that rely on the FB. This must be checked to be sure.