@martinwhitaker 2025-02-09
Personally I find `SCREEN_WIDTH - 1` just as easy to read and understand as `SCREEN_MAX_COL`.

IMO, if you can see visible flickering that is reduced by using `printf_region()`, that is adequate justification for making the change. I've never noticed it myself, but I don't usually pay much attention to the screen when tests are running.