@01e3 2025-02-09
I can drop the SCREEN_MAX_COL change if this is preferred.

We don't have to introduce print{f,s}_region, so this change can be also dropped.

Instead of:
```
#define display_test_pattern_value(pattern) \
    printf_region(5, 39, SCREEN_MAX_COL, "0x%0*x", TESTWORD_DIGITS, pattern)
``` 
... we would need to keep using something like for every place and live with the need to write the row (`5` in this case) written twice:
```
#define display_test_pattern_value(pattern) \
    clear_screen_region(5, printf(5, 39, "0x%0*x", TESTWORD_DIGITS, pattern), 5, SCREEN_WIDTH - 1);
```

It still prevents the clear / write pattern plus is IMO still better than what we have right now. with row repeated three times and column two times :
```
#define display_test_pattern_value(pattern) \
    { \
        clear_screen_region(5, 39, 5, SCREEN_WIDTH - 1); \
        printf(5, 39, "0x%0*x", TESTWORD_DIGITS, pattern); \
    }
```
Please let me know.