@01e3 2025-02-10
Alright, I tested my HP Z440 system with 256GB RAM and E5-2667v4 (8C, 16T) by looping test #1. Memtest86+ booted in UEFI mode to get FB.

Without the changes in this PR, I get 8 passes in 60 seconds. The flickering effect is easily visible:
![obraz](https://github.com/user-attachments/assets/6f013106-52a2-4434-b9bf-e422df3a968b)

With the PR, I can get 24 runs in 60 seconds, so 3(!) times more. For completeness, I can re-do the test in BIOS (text) mode but I expect the result to be similar. 

The flickering effect is now also gone, but due to the frequency of updates we still get some character mixing:
![obraz](https://github.com/user-attachments/assets/cb040070-8f6b-4404-b35d-f7666d2f4d99)

As predicted, most of the CPU cycles were going to the FB handling. I would even say we could consider setting test #1 back to SEQ - the cost of running this test is now 1/3 of what it used to be. Here it seems to be about 2.5s per 256GB RAM per thread, and this is a pretty old system. I will re-enable SEQ and measure the time needed, but if that rule is correct, it would expect 40 seconds vs 2 minutes.

Later this week, I can re-assembly my E5-2697Av4 (16 cores, 32 threads) with 512GB RAM system (the best what I have on hand in the place I'm right now) and do some more benchmarks.